Move validation to backend & refactor
