--- packages/cli/lib/validators/__tests__/validatorTestUtils.js
@@ -1,8 +1,0 @@
-//HACK so we can keep this util file next to the tests that use it
-test.skip('skip', () => null);
-
-const MODULE_PATH = 'module/path';
-
-module.exports = {
-  MODULE_PATH,
-};
