--- packages/cli/lib/validators/index.js
@@ -1,8 +1,0 @@
-const ModuleValidator = require('./marketplaceValidators/module/ModuleValidator');
-const ModuleDependencyValidator = require('./marketplaceValidators/module/ModuleDependencyValidator');
-
-const MARKETPLACE_VALIDATORS = {
-  module: [ModuleValidator, ModuleDependencyValidator],
-};
-
-module.exports = MARKETPLACE_VALIDATORS;
